Move RendererGroup to a separate module
