+++ bokehjs/src/lib/models/renderers/index.ts
@@ -1,5 +1,6 @@
 export {ContourRenderer} from "./contour_renderer"
-export {GlyphRenderer} from "./glyph_renderer"
-export {GraphRenderer} from "./graph_renderer"
-export {GuideRenderer} from "./guide_renderer"
-export {Renderer, RendererGroup}      from "./renderer"
+export {GlyphRenderer}   from "./glyph_renderer"
+export {GraphRenderer}   from "./graph_renderer"
+export {GuideRenderer}   from "./guide_renderer"
+export {Renderer}        from "./renderer"
+export {RendererGroup}   from "./renderer_group"
